fix: Move utils to root
