+++ usagelogger/aiohttp.py
@@ -6,7 +6,7 @@ from aiohttp import web
 from usagelogger import HttpLogger, HttpMessage
 from usagelogger.http_request_impl import HttpRequestImpl
 from usagelogger.http_response_impl import HttpResponseImpl
-from usagelogger.utils.multipart_utils import decode_multipart
+from usagelogger.multipart_utils import decode_multipart
 
 
 def HttpLoggerForAIOHTTP(url: Optional[str] = None, rules: Optional[str] = None):
